// frontend/src/app/learncoreservices/models/page-student-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PageableObject } from '../models/pageable-object';
import { SortObject } from '../models/sort-object';
import { StudentResponse } from '../models/student-response';
export interface PageStudentResponse {
  content?: Array<StudentResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}